genereate train data from unsatcores from multiple folders

+++ src/train_data_collection/generate_train_data_from_unsatcores.py
@@ -19,50 +19,56 @@ from src.solver.utils_parser import perse_eq_file
 import os
 
 def main():
-    benchmark_name="unsatcores_01_track_multi_word_equations_eq_5_20_generated_train_1_5000_one_core"
-    folder="divided_1"
-    all_eq_folder=f"{bench_folder}/{benchmark_name}/{folder}/UNSAT"
-    train_folder=f"{bench_folder}/{benchmark_name}/{folder}/train"
-    create_folder(train_folder)
+    benchmark_name="unsatcores_01_track_multi_word_equations_eq_5_20_generated_train_1_10000_one_core"
+    for fo in os.listdir(bench_folder + "/" + benchmark_name):
+        folder_basename=os.path.basename(fo)
+        if "divided" in folder_basename:
+            folder=folder_basename
 
-    for eq_file in glob.glob(f"{all_eq_folder}/*.eq"):
-        file_name=strip_file_name_suffix(eq_file)
-        unsatcore_file=f"{file_name}.unsatcore"
+            #folder="divided_2"
 
-        parsed_eq = perse_eq_file(eq_file)["equation_list"]
-        parsed_unsatcore = perse_eq_file(unsatcore_file)["equation_list"]
+            all_eq_folder=f"{bench_folder}/{benchmark_name}/{folder}/UNSAT"
+            train_folder=f"{bench_folder}/{benchmark_name}/{folder}/train"
+            create_folder(train_folder)
 
-        branch_eq_satisfiability_list = []
-        for eq in parsed_eq:
-            if eq in parsed_unsatcore:
-                branch_eq_satisfiability_list.append((eq,UNSAT,eq.term_length))
-            else:
-                branch_eq_satisfiability_list.append((eq, UNKNOWN, eq.term_length))
+            for eq_file in glob.glob(f"{all_eq_folder}/*.eq"):
+                file_name=strip_file_name_suffix(eq_file)
+                unsatcore_file=f"{file_name}.unsatcore"
 
-        node_id=1
-        label_list = [0] * len(branch_eq_satisfiability_list)
-        satisfiability_list = []
-        back_track_count_list = []
-        middle_branch_eq_file_name_list = []
-        one_train_data_name = f"{train_folder}/{os.path.basename(file_name)}@{node_id}"
-        for index, (eq, satisfiability, branch_number) in enumerate(branch_eq_satisfiability_list):
-            satisfiability_list.append(satisfiability)
-            back_track_count_list.append(branch_number)
-            one_eq_file_name = f"{one_train_data_name}:{index}"
+                parsed_eq = perse_eq_file(eq_file)["equation_list"]
+                parsed_unsatcore = perse_eq_file(unsatcore_file)["equation_list"]
 
-            eq.output_eq_file_rank(one_eq_file_name, satisfiability)
-            middle_branch_eq_file_name_list.append(os.path.basename(one_eq_file_name) + ".eq")
+                branch_eq_satisfiability_list = []
+                for eq in parsed_eq:
+                    if eq in parsed_unsatcore:
+                        branch_eq_satisfiability_list.append((eq,UNSAT,eq.term_length))
+                    else:
+                        branch_eq_satisfiability_list.append((eq, UNKNOWN, eq.term_length))
 
-        get_unsat_label(satisfiability_list,label_list,back_track_count_list)
+                node_id=1
+                label_list = [0] * len(branch_eq_satisfiability_list)
+                satisfiability_list = []
+                back_track_count_list = []
+                middle_branch_eq_file_name_list = []
+                one_train_data_name = f"{train_folder}/{os.path.basename(file_name)}@{node_id}"
+                for index, (eq, satisfiability, branch_number) in enumerate(branch_eq_satisfiability_list):
+                    satisfiability_list.append(satisfiability)
+                    back_track_count_list.append(branch_number)
+                    one_eq_file_name = f"{one_train_data_name}:{index}"
 
-        label_dict = {"satisfiability_list": satisfiability_list, "back_track_count_list": back_track_count_list,
-                      "label_list": label_list, "middle_branch_eq_file_name_list": middle_branch_eq_file_name_list}
-        dump_to_json_with_format(label_dict, one_train_data_name + ".label.json")
-        copy_relative_files(prefix=file_name,target_folder=train_folder)
+                    eq.output_eq_file_rank(one_eq_file_name, satisfiability)
+                    middle_branch_eq_file_name_list.append(os.path.basename(one_eq_file_name) + ".eq")
 
-    #zip train folder
-    zip_folder(folder_path=train_folder, output_zip_file=train_folder + ".zip")
-    shutil.rmtree(train_folder)
+                get_unsat_label(satisfiability_list,label_list,back_track_count_list)
+
+                label_dict = {"satisfiability_list": satisfiability_list, "back_track_count_list": back_track_count_list,
+                              "label_list": label_list, "middle_branch_eq_file_name_list": middle_branch_eq_file_name_list}
+                dump_to_json_with_format(label_dict, one_train_data_name + ".label.json")
+                copy_relative_files(prefix=file_name,target_folder=train_folder)
+
+            #zip train folder
+            zip_folder(folder_path=train_folder, output_zip_file=train_folder + ".zip")
+            shutil.rmtree(train_folder)
